Create Web_Scraping From Noon and Jumia sites
This project is a Python-based web scraping tool that extracts product data from Noon and Jumia,It integrates with the Gemini API to enhance data processing and analysis. A user-friendly GUI is built using CustomTkinter, making it easy to run scraping tasks, view logs, and interact with the tool without using the command line.